Keep working on experiment logic 2
