+++ experiments/pure_mcts/create_experiments_from_config.py
@@ -1,0 +1,147 @@
+import matplotlib.pyplot as plt
+import os
+import pandas as pd
+import seaborn as sns
+import shlex
+import shutil
+import subprocess
+import sys
+import yaml
+
+
+default_number_of_games = 100
+experiments_directory = '/shared/polygames-parent/experiments/pure_mcts'
+
+
+def get_directory_name_from_command(game_command, num_games, hyphenated: bool = False):
+    # Make a directory name unique to these hyperparameters
+    game_command = [el for el in game_command if el not in [
+        "python", "-m", "pypolygames", "pure_mcts"]]
+    dir_name = []
+    for item in game_command:
+        if not item.startswith("--"):
+            dir_name.append(item)
+    dir_name.append(str(num_games))
+    if hyphenated:
+        dir_name = '-'.join(dir_name)
+    else:
+        dir_name = '_'.join(dir_name)
+
+    return dir_name
+
+
+def get_experiment_commands(game_command, num_games):
+    all_experiment_commands = []
+
+    for i in range(num_games):
+        all_experiment_commands.append(game_command + ["--seed", str(i)])
+
+    return all_experiment_commands
+
+
+def run_games(num_games: int, save_plots: bool = True) -> tuple[list[str], list[str]]:
+
+    with open('config.yaml', 'r') as f:
+        hyperparameters = yaml.load(f, Loader=yaml.FullLoader)
+
+    for game_name in hyperparameters['game_name']:
+        for num_rollouts in hyperparameters['num_rollouts']:
+            for num_rollouts_2 in hyperparameters['num_rollouts_2']:
+             # Set all the hyperparameters, fixing the values of the ones
+             # we've already set
+                game_command = ["python", "-m", "pypolygames", "pure_mcts"]
+                for key, value in hyperparameters.items():
+                    if key == 'game_name':
+                        game_command += ["--" + key, game_name]
+                    elif key == 'num_rollouts':
+                        game_command += ["--" + key, str(num_rollouts)]
+                    elif key == 'num_rollouts_2':
+                        game_command += ["--" + key, str(num_rollouts_2)]
+                    else:
+                        game_command += ["--" + key, str(value)]
+
+                all_results = []
+                all_errors = []
+                experiment_name = get_directory_name_from_command(
+                    game_command, num_games, hyphenated=True)
+
+                all_experiment_commands = get_experiment_commands(game_command, num_games)
+                dir_name = get_directory_name_from_command(game_command, num_games)
+                print("saving yay!")
+
+                directory_path = f"{experiments_directory}/{dir_name}"
+
+                if os.path.exists(directory_path):
+                    print("WARNING: directory already exists. We're going to overwrite it.")
+                    shutil.rmtree(directory_path)
+                os.makedirs(directory_path)
+
+                with open(f"{directory_path}/game_command.txt", "w") as f:
+                    # First, write the number of games
+                    f.write(str(num_games) + '\n')
+
+                    # Then write the specific commands
+                    for experiment_command in all_experiment_commands:
+                        f.write(shlex.join(experiment_command) + '\n')
+
+                print("we just wrote the following file in the following directory")
+                print(f"{directory_path}/game_command.txt")
+
+                container = "http://ghcr.io/alignmentresearch/polygames:1.4.1-runner"
+
+                # os.system(f"python run_given_experiment.py {directory_path}/game_command.txt")
+
+                docker_command = f'ctl job run --name "nhowe-{experiment_name}" ' \
+                    f'--shared-host-dir-slow-tolerant --container "{container}" --cpu 4 --gpu 1 ' \
+                    '--login --never-restart --shared-host-dir /nas/ucb/k8 --shared-host-dir-mount /shared'\
+                    f'--command "python run_docker_experiments.py {directory_path}/game_command.txt"'
+                
+                print("the docker command is")
+                print(docker_command)
+                
+                raise SystemExit
+
+                # docker_command = f'ctl job run --name "nhowe-{experiment_name}" ' \
+                #     '--shared-host-dir-slow-tolerant --container "$CONTAINER" --cpu 4 --gpu 1 ' \
+                #     '--login --wandb --never-restart --shared-host-dir /nas/ucb/k8 --shared-host-dir-mount /shared'\
+                #     f'--command "python run_docker_experiments.py {directory_path}/game_command.txt"'
+
+                # Run the docker command
+                print("running the following docker command")
+                subprocess.run(shlex.split(docker_command))
+
+
+    # TODO: make it so it doesn't just return the most recent thing
+    return all_results, all_errors, game_command, dir_name
+
+
+if __name__ == "__main__":
+    # If filenames are passed in as arguments, use those. Otherwise, generate.
+    if len(sys.argv) == 2:
+        print("Checking if it's a path or a number of games...")
+        if os.path.isdir(sys.argv[1]):
+            print("Trying to load games from files...")
+            dir_name = sys.argv[1]
+            all_results, all_errors, game_command = get_results_from_directory(
+                dir_name)
+        elif sys.argv[1].isdigit():
+            print(f"Running {sys.argv[1]} games...")
+            all_results, all_errors, game_command, dir_name = run_games(
+                sys.argv[1])
+    else:
+        print(
+            f"Running default number of games ({default_number_of_games})...")
+        all_results, all_errors, game_command, dir_name = run_games(
+            default_number_of_games)
+
+    print("all the results and errors are")
+    print(all_results)
+    print(all_errors)
+
+    print("the game command is")
+    print(game_command)
+
+    # Now make the plot
+    print("Making plot...")
+    print("the game command was", game_command)
+    make_plot(all_results, all_errors, game_command, dir_name)
